1. Fixed a small issues with the logging call in Schedulers.jl 2. Moved all logging and related funcitons to Log.jl 3. Updated "using" statements in Scheduler.jl and Log.jl 4. Orginal logging code is left commented out. 5. Updated exports of Log.jl to export the logging functions
